Misc PR adjustments; Resolve translation files + file structure cleanup.
